+++ test/Braintree.Tests/UsBankAccountVerificationTest.cs
@@ -28,9 +28,11 @@ namespace Braintree.Tests
             builder.Append("<?xml version=\"1.0\" encoding=\"UTF-8\"?>");
             builder.Append("<thing>");
             builder.Append("  <verification>");
-            builder.Append("    <verification-method>independent_check</verification-method>");
+            builder.Append("    <verification-method>network_check</verification-method>");
+            builder.Append("    <verification-add-ons>customer_verification</verification-add-ons>");
             builder.Append("    <status>processor_declined</status>");
             builder.Append("    <processor-response-code>2000</processor-response-code>");
+            builder.Append("    <additional-processor-response>Invalid routing number</additional-processor-response>");
             builder.Append("    <processor-response-text>Do Not Honor</processor-response-text>");
             builder.Append("  </verification>");
             builder.Append("</thing>");
@@ -42,10 +44,12 @@ namespace Braintree.Tests
                 new NodeWrapper(doc).GetNode("//verification")
             );
 
-            Assert.AreEqual(UsBankAccountVerificationMethod.INDEPENDENT_CHECK, verification.VerificationMethod);
+            Assert.AreEqual(UsBankAccountVerificationMethod.NETWORK_CHECK, verification.VerificationMethod);
+            Assert.AreEqual(VerificationAddOns.CUSTOMER_VERIFICATION, verification.VerificationAddOns);
             Assert.AreEqual(UsBankAccountVerificationStatus.PROCESSOR_DECLINED, verification.Status);
             Assert.AreEqual("2000", verification.ProcessorResponseCode);
             Assert.AreEqual("Do Not Honor", verification.ProcessorResponseText);
+            Assert.AreEqual("Invalid routing number", verification.AdditionalProcessorResponse);
         }
 
         [Test]