+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## 5.21.0
+- Add `DateOfBirth` and `CountryCode` to `IndustryDataRequest`
+- Add `BillingAddress` and `ShippingAddress` to `VenmoProfileData`
+- Add `VerificationAddOns` and `AdditionalProcessorResponse` to `UsBankAccountVerification`
+- Add `MetaCheckoutCard` and `MetaCheckoutToken` to payment method
+- Add `MetaCheckoutCardDetails` nd `MetaCheckoutTokenDetails` to Transaction
+
 ## 5.20.0
 - Add `SUBSCRIPTIONBILLINGSKIPPED` webhook notification support
 - Add `implicitlyVaultedPaymentMethodToken` and `implicitlyVaultedPaymentMethodGlobalId` to `LocalPaymentDetails`